Refactor initialize in TDDTestKeeper
The new implementation will try to look for `InjectionIII` and `injectionforxcode`, the first path that can be resolved will be used.
InjectionIII takes precedence of injectionforxcode because that version of Injection for Xcode is newer and is currently being
maintained. In addition to loading InjectionIII bundle, it will also try and resolve the different versions of the bundle, the first
one being the one that is built for Xcode 10 and if that one cannot be resolved it will try and load the Xcode 9 fallback version.